mini-server-prod/kgroot-services/requirements.txt
# KGroot Implementation Requirements

# Core dependencies
networkx==3.2.1
neo4j==5.14.0
numpy==1.26.2
pandas==2.1.4

# GraphRAG and AI
openai==1.6.1
tiktoken==0.5.2
langchain==0.1.0
langchain-openai==0.0.2

# Vector storage
chromadb==0.4.22
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3
python-dateutil==2.8.2

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0